Convert to use Postgres, add events and help pages
* Switch database connection to be heroku env var compatible
* Set DB logging to STDOUT
* Add events and help page

+++ src/models/event.cr
@@ -9,7 +9,7 @@ class Event < Crecto::Model
   end
 
   schema "events" do
-    field :object_id, Int64
+    field :object_id, Int32
     field :object_class, String
     enum_field :action, Action, column_name: "action_type", column_type: Int32
   end
